+++ app/controls/PanelBodyWithAdd.js
