+++ app/controls/PanelBodyWithToggle.js
